feat(Gadget-libCachedCode): Add caching mechanism and support via `libCachedCode`
Add caching mechanism for `libCachedCode` to support retrieving cached code and injecting it to the page, with support for `script` and `css` types. Also, add `libCachedCode` as a readonly global variable and a hidden gadget in `.eslintrc.yaml`, `settings.json`, and `definition.yaml`. The `MediaWiki:Gadget-libCachedCode.js` script provides methods to get cached code, obtain the URL for cached code and inject cached code to the page.