feat: Add codemirrorMediawiki method for highlighting and linting wikitext
This commit adds a codemirrorMediawiki method to the wikieditor-highlight gadget, which can be used to highlight and lint wikitext. The method is used to configure the CodeMirror instance to support the syntax for MediaWiki articles and its extension tags. The change also updates the Gadget-wikieditor-highlight.js file to use the new method.