feat(prefetch): Add fetchNPMPackageInfo to get package info
This code change introduces the fetchNPMPackageInfo module to the prefetch script. The module is used to fetch the package info from the registry instead of using the exec module to get the registry base URL and running a fetch command. The resulting package info allows for the retrieval of the correct package version based on the semver expression.