feat(browserify,mkdtmp): Add new feature of creating a local tmp file
This commit adds a new feature to create a local tmp file in the `mkdtmp` module. Instead of using the default location, the new feature uses a local folder named `.tmp` to create a random or fixed file based on the input. This change is then utilized by the `browserify` script to create a temporary path for the input file.