feat(ci): Backup and restore package-lock.json file
Add functionality to backup and restore package-lock.json file before and after a job pipeline, respectively. Now before running the pipeline, package-lock.json is backed up to a temporary directory, and during recovering, the backup file is retrieved first and then the file is recovered from git. The backuped file is used if it exists otherwise git is used to restore the file.